Break `scripts.js` into smaller files
