+++ package.json
@@ -1,5 +1,7 @@
 {
   "devDependencies": {
+    "@babel/core": "^7.14.6",
+    "@babel/preset-env": "^7.14.7",
     "autoprefixer": "^10.2.6",
     "browser-sync": "^2.27.4",
     "del": "^6.0.0",
@@ -13,6 +15,7 @@
     "eslint-plugin-security": "^1.4.0",
     "eslint-plugin-sort-keys-fix": "^1.1.1",
     "gulp": "^4.0.2",
+    "gulp-babel": "^8.0.0",
     "gulp-concat": "^2.6.1",
     "gulp-dart-sass": "^1.0.2",
     "gulp-postcss": "^9.0.0",
@@ -20,6 +23,7 @@
     "gulp-uglify": "^3.0.2",
     "gulp-zip": "^5.1.0",
     "js-yaml": "^4.1.0",
+    "merge2": "^1.4.1",
     "node-fetch": "^2.1.2",
     "npm-run-all": "^4.1.5",
     "sass": "^1.35.1",
