+++ src/scripts/pages/new-community-post-page.js
@@ -1,0 +1,10 @@
+document.addEventListener('DOMContentLoaded', function() {
+  var notificationElm = document.querySelector(".notification-error");
+  if (
+    notificationElm &&
+    notificationElm.previousElementSibling &&
+    typeof notificationElm.previousElementSibling.focus === "function"
+  ) {
+    notificationElm.previousElementSibling.focus();
+  }
+});
