+++ src/scripts/pages/requests.js
@@ -1,0 +1,46 @@
+document.addEventListener('DOMContentLoaded', () => {
+  var ENTER = 13;
+
+  function saveFocus() {
+    var activeElementId = document.activeElement.getAttribute("id");
+    sessionStorage.setItem('returnFocusTo', '#' + activeElementId);
+  }
+
+  function restoreFocus() {
+    var returnFocusTo = sessionStorage.getItem('returnFocusTo');
+    if (returnFocusTo) {
+      sessionStorage.removeItem('returnFocusTo');
+      var returnFocusToEl = document.querySelector(returnFocusTo);
+      returnFocusToEl && returnFocusToEl.focus && returnFocusToEl.focus();
+    }
+  }
+
+  restoreFocus();
+
+  // Submit requests filter form on status or organization change in the request list page
+  Array.prototype.forEach.call(document.querySelectorAll('#request-status-select, #request-organization-select'), function(el) {
+    el.addEventListener('change', function(e) {
+      e.stopPropagation();
+      saveFocus();
+      closest(this, 'form').submit();
+    });
+  });
+
+  // Submit requests filter form on search in the request list page
+  var quickSearch = document.querySelector('#quick-search');
+  quickSearch && quickSearch.addEventListener('keyup', function(e) {
+    if (e.keyCode === ENTER) {
+      e.stopPropagation();
+      saveFocus();
+      closest(this, 'form').submit();
+    }
+  });
+
+  // Submit organization form in the request page
+  var requestOrganisationSelect = document.querySelector('#request-organization select');
+  if (requestOrganisationSelect) {
+    requestOrganisationSelect.addEventListener('change', function() {
+      closest(this, 'form').submit();
+    });
+  }
+});
