+++ src/scripts/ui/collapsible-nav.js
@@ -1,0 +1,32 @@
+document.addEventListener('DOMContentLoaded', () => {
+  var ESCAPE = 27;
+
+  function toggleNavigation(toggle, menu) {
+    var isExpanded = menu.getAttribute('aria-expanded') === 'true';
+    menu.setAttribute('aria-expanded', !isExpanded);
+    toggle.setAttribute('aria-expanded', !isExpanded);
+  }
+
+  function closeNavigation(toggle, menu) {
+    menu.setAttribute('aria-expanded', false);
+    toggle.setAttribute('aria-expanded', false);
+    toggle.focus();
+  }
+
+  // Toggles expanded aria to collapsible elements
+  var collapsible = document.querySelectorAll('.collapsible-nav, .collapsible-sidebar');
+
+  Array.prototype.forEach.call(collapsible, function(el) {
+    var toggle = el.querySelector('.collapsible-nav-toggle, .collapsible-sidebar-toggle');
+
+    el.addEventListener('click', function(e) {
+      toggleNavigation(toggle, this);
+    });
+
+    el.addEventListener('keyup', function(e) {
+      if (e.keyCode === ESCAPE) {
+        closeNavigation(toggle, this);
+      }
+    });
+  });
+});
