+++ src/scripts/ui/filter-list.js
@@ -1,0 +1,18 @@
+document.addEventListener('DOMContentLoaded', () => {
+  // If multibrand search has more than 5 help centers or categories collapse the list
+  var multibrandFilterLists = document.querySelectorAll(".multibrand-filter-list");
+  Array.prototype.forEach.call(multibrandFilterLists, function(filter) {
+    if (filter.children.length > 6) {
+      // Display the show more button
+      var trigger = filter.querySelector(".see-all-filters");
+      trigger.setAttribute("aria-hidden", false);
+
+      // Add event handler for click
+      trigger.addEventListener("click", function(e) {
+        e.stopPropagation();
+        trigger.parentNode.removeChild(trigger);
+        filter.classList.remove("multibrand-filter-list--collapsed")
+      })
+    }
+  });
+});
