+++ src/scripts/ui/navigation.js
@@ -1,0 +1,8 @@
+document.addEventListener('DOMContentLoaded', () => {
+  const $headerCheckbox = document.getElementById('header-checkbox');
+
+  $headerCheckbox.addEventListener('click', () => {
+    $headerCheckbox.classList.toggle('closed');
+    $headerCheckbox.classList.toggle('open');
+  });
+});
