+++ src/scripts/ui/social-sharing.js
@@ -1,0 +1,8 @@
+document.addEventListener('DOMContentLoaded', () => {
+  Array.prototype.forEach.call(document.querySelectorAll('.share a'), function(anchor) {
+    anchor.addEventListener('click', function(e) {
+      e.preventDefault();
+      window.open(this.href, '', 'height = 500, width = 500');
+    });
+  });
+});
